% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{check_system_library}
\alias{check_system_library}
\title{Check system library for unallwed packages}
\usage{
check_system_library(allowed = getOption("rprofile.system_library",
  c("remotes", "rprofile")))
}
\arguments{
\item{allowed}{Character vector of allowed packages.
Defaults to "rprofile.system_library" user option if set, else
`c("remotes", "rprofile")`}
}
\description{
Check system library for unallwed packages
}
