% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-projects.R
\name{cp}
\alias{cp}
\title{Choose an RStudio Project}
\usage{
cp(path = NULL)
}
\arguments{
\item{path}{Default \code{NULL}. If not \code{NULL}, path is passed to \code{op}.}
}
\description{
Command line version for choosing RStudio projects
}
\examples{
\dontrun{
cp()
}
}
