% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_generator.R
\name{create_make_functions}
\alias{create_make_functions}
\title{Generate make functions

Parses Makefile and automatically creates make_* functions}
\usage{
create_make_functions()
}
\description{
Generate make functions

Parses Makefile and automatically creates make_* functions
}
