% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{.Air}
\alias{.Air}
\title{Run air}
\usage{
.Air(cmd = c("format", "language-server", "help"), ...)
}
\arguments{
\item{cmd}{The command to run (not the COMMAND option for \verb{air help <COMMAND>})}

\item{...}{Additional arguments to pass to the command.}
}
\description{
Wrapper for calling \code{air}
}
\examples{
\dontrun{
.Air("format", paths = "R", check = TRUE, log_level = "trace")
}
}
