% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cran-status.R
\name{.CheckCranStatus}
\alias{.CheckCranStatus}
\title{Check Cran Status}
\usage{
.CheckCranStatus(email = NULL)
}
\arguments{
\item{email}{The email address to query CRAN}
}
\description{
A wrapper
}
\seealso{
\link[dang:checkCRANStatus]{dang::checkCRANStatus}
}
