% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{.FindRprofile}
\alias{.FindRprofile}
\title{Find \code{.Rprofile}}
\usage{
.FindRprofile(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{TRUE} silences messages, otherwise provides information
about the location of \code{.Rprofile}}
}
\value{
See \strong{Details}
}
\description{
Find \code{.Rprofile}
}
\details{
The first location searched is
}
