% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{.LintFile}
\alias{.LintFile}
\title{Lint a single file}
\usage{
.LintFile(path = NULL)
}
\arguments{
\item{path}{The file path.  Default (\code{NULL}) checks for the current file open
in the source editor in \strong{RStudio}.}
}
\value{
See \code{lintr::lint_dir()}
}
\description{
Use \code{lintr::lint_dir()} to \emph{lint} a single file
}
