% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{.NewsUrls}
\alias{.NewsUrls}
\title{Add URLs to NEWS}
\usage{
.NewsUrls(path = ".", ask = interactive(), url = NULL, check)
}
\arguments{
\item{path}{The path to your package}

\item{ask}{Prompts updating \code{NEWS.md}}

\item{url}{Optional URL to set for bug reports}
}
\value{
Nothing, called for its side effects
}
\description{
Include links to github on \code{NEWS.md}
}
