% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-url.R
\name{.OpenPackageUrl}
\alias{.OpenPackageUrl}
\title{Open package url}
\usage{
.OpenPackageUrl(pkg)
}
\arguments{
\item{pkg}{A character vector of package names}
}
\value{
Nothing, used for its side-effects
}
\description{
Open package url
}
