% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak.R
\name{.Pak}
\alias{.Pak}
\title{Wrapper for \code{{pak::pak}}}
\usage{
.Pak(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[pak:pak]{pak::pak()}}.}
}
\description{
Wrapper for \code{{pak::pak}}
}
\details{
If the envvar \code{R_LIBS_PAK} is set then it will be used as the
library path for \code{{pak}}
}
