% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{.ReadClip}
\alias{.ReadClip}
\title{Read the clipboard}
\usage{
.ReadClip(convert = c("auto", "none", "some", "data.frame", "tibble"), ...)
}
\arguments{
\item{convert}{The method to use.
\itemize{
\item \code{"auto"}: (default) Parses contents either into a vector and applies some
conversion; if multiple lines are detected, attempts to parse as a table
\item \code{"none"}: Returns the contents as a string
\item \code{"some"}: Returns the contents as a character vector and performs some
parsing to possibly convert to a vector
\item \code{"data.frame"}: Returns the contents as a data frame (via
\code{clipr::read_clip_tbl()})
}}

\item{...}{Additional arguments passed to \code{\link[utils:read.table]{utils::read.table()}} (defaults
have been changed)}
}
\description{
Reads the clipboard and returns the contents as a character vector or table
}
\details{
Essentially wraps \code{clipr::read_clip()}
}
