% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{.ResetOptions}
\alias{.ResetOptions}
\title{Reset options}
\usage{
.ResetOptions(keep_prompt = TRUE)
}
\arguments{
\item{keep_prompt}{Logical, if \code{FALSE} will not reset the \code{prompt} option}
}
\description{
Reset options
}
