% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{.Rprofile}
\alias{.Rprofile}
\title{Get .Rprofile}
\usage{
.Rprofile(update = FALSE, overwrite = FALSE)
}
\arguments{
\item{update}{If \code{TRUE} will copy the .Rprofile from the Rprofile package}

\item{overwrite}{if attempting to update, also overwrite}
}
\description{
Get .Rprofile
}
