% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofile-package.R, R/rprofile.R
\name{.RprofileJordan}
\alias{.RprofileJordan}
\title{Jordan's Rprofile options}
\usage{
.RprofileJordan()

.RprofileJordan()
}
\description{
Jordan's Rprofile options

Jordan's Rprofile options
}
