% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{.SpellCheckFile}
\alias{.SpellCheckFile}
\title{Spell check a file}
\usage{
.SpellCheckFile(path = NULL)
}
\arguments{
\item{path}{A path to spell check; if \code{NULL} check the current}
}
\description{
Spell check a file
}
