% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-actions.R
\name{.UseGithubAction}
\alias{.UseGithubAction}
\title{Use GitHub Action}
\usage{
.UseGithubAction(x, repo = NULL, name = NULL, ref = NULL, ask = interactive())
}
\arguments{
\item{x, repo, name, ref}{Name/location of the action.  You can use a format
such as \code{"<<repo>|<owner>/<repo>>/<[examples/]/<path>>@<ref>"}.  This
function will automatically check an \verb{examples/} subdirectory in the
repository.}

\item{ask}{Checks if existing files should be overwritten}
}
\description{
Use GitHub Action
}
\examples{
\dontrun{
.UseGithubAction("jmbarbone/actions/r-check-version")
.UseGithubAction("r-lib/actions/lint@v2")
}

}
