% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-terminal.R
\name{is_terminal}
\alias{is_terminal}
\alias{set_terminal}
\alias{set_rstudio}
\title{Detects R terminal}
\usage{
is_terminal()

set_terminal(rdoc = TRUE, colorout = TRUE, prettycode = TRUE)

set_rstudio(prettycode = TRUE)
}
\arguments{
\item{rdoc}{Should we load the rdoc package}

\item{colorout}{Should we load the colorout package}

\item{prettycode}{Should we load the prettycode package}
}
\description{
Return \code{TRUE} is terminal. But \code{FALSE} if the RStudio console.
The RStudio console doesn't require as much customisation as a standard console since
it is already pretty. But it does place limits on the prompt options (single line only).
}
