% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_style.R
\name{jr_style}
\alias{jr_style}
\title{R Code styling}
\usage{
jr_style(scope = "tokens")
}
\arguments{
\item{scope}{The extent of manipulation. Can range from "none" (least
invasive) to "tokens" (most invasive). See 'Details'. This argument is a
string or a vector of class \code{AsIs}.}
}
\description{
R Code styling
}
