% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{library_name}
\alias{library_name}
\title{Library name}
\usage{
library_name(cran = getOption("rprofile.cran", version_cran()),
  bioc = getOption("rprofile.bioc", version_bioc()))
}
\arguments{
\item{cran}{The CRAN version.
Defaults to the "rprofile.cran" option if set, else
\code{\link[rprofile:version_cran]{rprofile::version_cran()}}.}

\item{bioc}{The Bioconductor version.
Defaults to the "rprofile.bioc" option if set, else
\code{\link[rprofile:version_bioc]{rprofile::version_bioc()}}.}
}
\value{
Returns a string which will be used as the versioned library name.
}
\description{
Construct a library name from R, Bioconductor, and CRAN version information.
}
