% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{library_path}
\alias{library_path}
\title{Construct the library path}
\usage{
library_path(home = getOption("rprofile.lib.home", R.home()),
  name = getOption("rprofile.lib.name", library_name()))
}
\arguments{
\item{home}{Path to home of versioned library.
Defaults to the "rprofile.lib.home" option if set, else
\code{\link[base:R.home]{base::R.home()}}}

\item{name}{Name of versioned library.
Defaults to the "rprofile.lib.name" option if set, else
\code{\link[rprofile:library_name]{rprofile::library_name()}}}
}
\description{
Constructs a path to the versioned library in the intended installation home.
}
