% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-projects.R
\name{op}
\alias{op}
\title{RStudio projects

A command line version of opening RStudio projects.}
\usage{
op(path = ".")
}
\arguments{
\item{path}{Path to the (proposed) RStudio project.}
}
\description{
RStudio projects

A command line version of opening RStudio projects.
}
\examples{
\dontrun{
# Open project in current working directory
op()
# Open project in current working directory
op("/path/to/project")
}
}
