% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-file.R
\name{OpenFile}
\alias{OpenFile}
\alias{.OpenFile}
\alias{.OpenFile.default}
\alias{.OpenFile.data.frame}
\alias{.OpenFile.matrix}
\title{Open (as) File}
\usage{
.OpenFile(x, ...)

\method{.OpenFile}{default}(x, force = FALSE, ...)

\method{.OpenFile}{data.frame}(x, ...)

\method{.OpenFile}{matrix}(x, ...)
}
\arguments{
\item{x}{A object.  If an existing file path, runs with \code{xopen::xopen}.}

\item{...}{Not used}

\item{force}{If \code{TRUE} ignores potential file path in \code{x}}
}
\value{
The path to the open file
}
\description{
Open a file path or data as a file
}
