% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{Reload}
\alias{Reload}
\alias{.Reload}
\alias{.Restart}
\title{Reload}
\usage{
.Reload(remove_objects = TRUE, loud = FALSE)

.Restart()
}
\arguments{
\item{remove_objects}{Logical, if \code{TRUE} will remove all objects found}

\item{loud}{Logical, if \code{TRUE} will present a message on removed objects}
}
\description{
Reloads your session
}
