% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{remove_tag_and_save}
\alias{remove_tag_and_save}
\title{Finds where tag is in file and removes the line and the line after
Assumes that the 'tag' is from this package and starts with
'@jordan_'
setting "tag" to NULL will remove any instances}
\usage{
remove_tag_and_save(file, tag, warn = TRUE)
}
\arguments{
\item{file}{A file path}

\item{tag}{A character tag}

\item{warn}{if \code{FALSE} will not produce a warning message}
}
\description{
Finds where tag is in file and removes the line and the line after
Assumes that the 'tag' is from this package and starts with
'@jordan_'
setting "tag" to NULL will remove any instances
}
