% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{repos}
\alias{repos}
\title{Create repository vector}
\usage{
repos(..., cran = version_cran(), bioc = version_bioc(),
  cran_mirror = getOption("rprofile.cran.mirror",
  "https://cran.microsoft.com/snapshot"),
  bioc_mirror = getOption("rprofile.bioc.mirror", "https://bioconductor.org"))
}
\arguments{
\item{...}{User specified repositories added to beginning of vector.}

\item{cran}{The CRAN snapshot date ("YYYY-MM-DD").
Defaults to \code{rprofile::version_cran()}.}

\item{bioc}{The Bioconductor version.
Defaults to \code{rprofile::version_bioc()}.}

\item{cran_mirror}{URL for CRAN snapshot mirror.
Defaults to \code{getOption("rprofile.cran", "https://cran.microsoft.com/snapshot")}.}

\item{bioc_mirror}{URL for Bioconductor mirror.
Defaults to \code{getOption("rprofile.bioc", "https://bioconductor.org")}.}
}
\description{
Create a named vector of repositories
}
\examples{
rprofile::repos()

# Change the CRAN snapshot to an old date intended for a different R version
r <- "3.3"

rprofile::repos(
  my_repo = "url/to/CRAN/like/repo",
  cran    = version_cran(r)
)

}
