% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofile-package.R
\docType{package}
\name{Rprofile-package}
\alias{Rprofile}
\alias{Rprofile-package}
\title{Rprofile: Functions to be loaded with Rprofile}
\description{
Provides a set of non-intrusive development utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmbarbone/Rprofile}
  \item \url{https://jmbarbone.github.io/Rprofile/}
  \item Report bugs at \url{https://github.com/jmbarbone/Rprofile/issues}
}

}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]
}

}
\keyword{internal}
