% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{rprofile}
\alias{rprofile}
\title{Environment setup message}
\usage{
rprofile(verbose = interactive())
}
\arguments{
\item{verbose}{Logical. Should the message be issued. Defaults to \code{TRUE} in
an interactive session, \code{FALSE} otherwise.}
}
\description{
Message the user to indicate current system, repository and library
configuration.
}
