% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_callback<-}
\alias{rprofile_callback<-}
\title{Modify an R profile task callback.}
\usage{
rprofile_callback(callback, s3key) <- value
}
\arguments{
\item{callback}{character. The name of the task callback.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{The}{value to set.}
}
\value{
The character value of the task callback.
}
\description{
Modify an R profile task callback.
}

