% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_callback}
\alias{rprofile_callback}
\title{Access an R profile task callback.}
\usage{
rprofile_callback(callback, s3key, remove)
}
\arguments{
\item{callback}{character. The name of the task callback.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{remove}{logical. If not missing and set to \code{TRUE},
remove the \code{callback} instead.}
}
\value{
The character value of the task callback.
}
\description{
Access an R profile task callback.
}

