% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_callbacks}
\alias{rprofile_callbacks}
\title{List all Rprofile task callbacks.}
\usage{
rprofile_callbacks(callbacks)
}
\arguments{
\item{callbacks}{list. If not missing, replace callbacks instead.}
}
\value{
A named list of task callbacks.
}
\description{
List all Rprofile task callbacks.
}

