% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_code<-}
\alias{rprofile_code<-}
\title{Modify an R profile code section.}
\usage{
rprofile_code(code, s3key) <- value
}
\arguments{
\item{code}{character. The name of the code section.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{The}{value to set.}
}
\value{
The character value of the code section.
}
\description{
Modify an R profile code section.
}

