% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_code}
\alias{rprofile_code}
\title{Access an R profile code section.}
\usage{
rprofile_code(code, s3key, remove)
}
\arguments{
\item{code}{character. The name of the code section or a code expression.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{remove}{logical. If not missing and set to \code{TRUE},
remove the \code{code} instead.}
}
\value{
The character value of the code section.
}
\description{
Access an R profile code section.
}

