% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_codes}
\alias{rprofile_codes}
\title{List all Rprofile code sections.}
\usage{
rprofile_codes(codes)
}
\arguments{
\item{codes}{list. If not missing, replace codes instead.}
}
\value{
A named list of code sections.
}
\description{
List all Rprofile code sections.
}

