% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_envvar}
\alias{rprofile_envvar}
\title{Access an R profile environment variable.}
\usage{
rprofile_envvar(envvar, s3key, remove)
}
\arguments{
\item{envvar}{character. The name of the environment variable.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{remove}{logical. If not missing and set to \code{TRUE},
remove the \code{envvar} instead.}
}
\value{
The character value of the environment variable.
}
\description{
Access an R profile environment variable.
}

