% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_envvars}
\alias{rprofile_envvars}
\title{List all Rprofile environment variables.}
\usage{
rprofile_envvars(options)
}
\arguments{
\item{options}{list. If not missing, replace envvars instead.}
}
\value{
A named character vector of environment variables.
}
\description{
List all Rprofile environment variables.
}

