% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_option<-}
\alias{rprofile_option<-}
\title{Modify an R profile global option.}
\usage{
rprofile_option(option, s3key) <- value
}
\arguments{
\item{option}{character. The name of the global option.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{The}{value to set.}
}
\value{
The character value of the global option.
}
\description{
Modify an R profile global option.
}

