% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_option}
\alias{rprofile_option}
\title{Access an R profile global option.}
\usage{
rprofile_option(envvar, s3key, remove)
}
\arguments{
\item{envvar}{character. The name of the global option.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{remove}{logical. If not missing and set to \code{TRUE},
remove the \code{option} instead.}
}
\value{
The character value of the global option.
}
\description{
Access an R profile global option.
}

