% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_options}
\alias{rprofile_options}
\title{List all Rprofile global options.}
\usage{
rprofile_options(options)
}
\arguments{
\item{options}{list. If not missing, replace options instead.}
}
\value{
A named list of global options.
}
\description{
List all Rprofile global options.
}

