% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_package<-}
\alias{rprofile_package<-}
\title{Modify an R profile autoloaded package.}
\usage{
rprofile_package(package, s3key) <- value
}
\arguments{
\item{package}{character. The name of the autoloaded package.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{The}{value to set.}
}
\value{
The character value of the autoloaded package.
}
\description{
Modify an R profile autoloaded package.
}

