% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_package}
\alias{rprofile_package}
\title{Access an R profile autoloaded package.}
\usage{
rprofile_package(package, s3key, remove)
}
\arguments{
\item{package}{character. The name of the autoloaded package.}

\item{s3key}{character. The s3key in which the Rprofile is stored.
By default, \code{option("rprofile.s3key")}.}

\item{remove}{logical. If not missing and set to \code{TRUE},
remove the \code{package} instead.}
}
\value{
The character value of the autoloaded package.
}
\description{
Access an R profile autoloaded package.
}

