% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_interfaces.R
\name{rprofile_packages}
\alias{rprofile_packages}
\title{List all Rprofile autoloaded packages.}
\usage{
rprofile_packages(packages)
}
\arguments{
\item{packages}{list. If not missing, replace packages instead.}
}
\value{
A named list of autoloaded packages.
}
\description{
List all Rprofile autoloaded packages.
}

