% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rprofile_section.R
\name{rprofile_section}
\alias{rprofile_section}
\title{Create a new object recording an Rprofile section.}
\usage{
rprofile_section(type)
}
\arguments{
\item{type}{character. The type of the section.}
}
\value{
An \code{rprofile_section} object.
}
\description{
The possible section types are
\code{c("envvar", "option", "packages", "callback", "code")}.
}

