% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{set_library}
\alias{set_library}
\title{Set the library path}
\usage{
set_library(lib = library_path())
}
\arguments{
\item{lib}{Path to versioned library.
Defaults to \code{\link[rprofile:library_path]{rprofile::library_path()}}.}
}
\description{
Adds the versioned library to the top of the list of previously set
libraries. If the library directory does not exist, it will be created.
}
