% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repos.R
\name{set_repos}
\alias{set_repos}
\title{Set R repositories
A simple function for setting R repositories.}
\usage{
set_repos(CRAN = "https://cran.rstudio.com/", ...)
}
\arguments{
\item{CRAN}{Default is RStudio's CRAN repo}

\item{...}{Other repositories}
}
\description{
Set R repositories
A simple function for setting R repositories.
}
\examples{
\dontrun{
set_repos(repoX = "https://example.com")
}
}
