% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{set_repositories}
\alias{set_repositories}
\title{Set repositories}
\usage{
set_repositories(repos = repos())
}
\arguments{
\item{repos}{A named character vector of repositories.
Defaults to \code{\link[rprofile:repos]{rprofile::repos()}}.}
}
\description{
Convenience function for setting repositories.
}
