% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_options.R
\name{set_startup_options}
\alias{set_startup_options}
\title{Set Nice Startup Options

Sets nicer options. All arguments are passed to the \code{options} function.
This function also sets \code{ipck = TRUE} in \code{rc.settings}.}
\usage{
set_startup_options(
  digits = 4L,
  show.signif.stars = FALSE,
  useFancyQuotes = FALSE,
  width = 88L,
  Ncpus = max(1L, parallel::detectCores() - 1L),
  continue = " ",
  max.print = 100L,
  servr.daemon = TRUE,
  max = 10L,
  mc.cores = max(1L, parallel::detectCores() - 1L),
  error = "rlang",
  menu.graphics = FALSE,
  warnPartialMatchArgs = TRUE,
  scipen = 999L,
  HTTPUserAgent = sprintf("R/\%s R (\%s)", getRversion(), paste(getRversion(),
    R.version$platform, R.version$arch, R.version$os)),
  download.file.extra = sprintf("--header \\"User-Agent: R (\%s)\\"",
    paste(getRversion(), R.version$platform, R.version$arch, R.version$os)),
  ...
)
}
\arguments{
\item{digits}{Default \code{4}}

\item{show.signif.stars}{Default \code{FALSE}}

\item{useFancyQuotes}{Default \code{FALSE}}

\item{width}{Default \code{100}}

\item{Ncpus}{Default number of CPUs - 1. Used for parallel pkg installs.}

\item{continue}{Default blank space (remove the default +)}

\item{max.print}{Default 100 to avoid blow up}

\item{servr.daemon}{Default \code{TRUE}.  For xaringan presentations}

\item{max}{Default \code{10}. For List printing}

\item{mc.cores}{Default number of CPUs - 1. Used for parallel computing}

\item{error}{Default \code{rlang}. If \code{rlang} is installed, then error = rlang::entrace.}

\item{menu.graphics}{Default \code{FALSE}. Logical: should graphical menus be used if available?}

\item{warnPartialMatchArgs}{Default \code{TRUE}. Warn if using partial arguments.}

\item{scipen}{Default \code{999}. Always print out full numbers, i.e. not 1e2}

\item{HTTPUserAgent}{Used by RStudio Package Manager (RSPM).}

\item{download.file.extra}{Used by RSPM for curl/wget installs, e.g. Rscript.}

\item{...}{Other arguments passed to \code{options}.}
}
\description{
Set Nice Startup Options

Sets nicer options. All arguments are passed to the \code{options} function.
This function also sets \code{ipck = TRUE} in \code{rc.settings}.
}
