% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-message.R
\name{startup_message}
\alias{startup_message}
\title{Startup message}
\usage{
startup_message(verbose = interactive())
}
\arguments{
\item{verbose}{Logical. Should the message be issued. Defaults to TRUE in
an interactive session, FALSE otherwise.}
}
\description{
Message the user to indicate current system, repository and library
configuration.
}
