% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-checker.R
\name{test_checker}
\alias{test_checker}
\title{Compare R files with test files}
\usage{
test_checker()
}
\description{
testthat has the naming convention of test-filename.R
This is a simple function that determines if any files
don't follow that convention.
It's meant to be a guide, not a perfect solution
}
