% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rprofile.R
\name{use_rprofile}
\alias{use_rprofile}
\title{Use rprofile}
\usage{
use_rprofile(dir = ".", template = system.file("templates/Rprofile.R",
  package = "rprofile"))
}
\arguments{
\item{dir}{Path to directory to create the Rprofile.
Defaults to the current working directory (i.e. ".").}

\item{template}{Path to Rprofile template.
Defaults to a template included in the rprofile package.}
}
\description{
Create an Rprofile from template.
}
