% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{version_bioc}
\alias{version_bioc}
\alias{map_r_to_bioc}
\title{Get Bioconductor version}
\usage{
version_bioc(r = version_r_major_minor(), map = map_r_to_bioc())

map_r_to_bioc(...)
}
\arguments{
\item{r}{The R version (e.g. "3.5").
Defaults to \code{version_r_major_minor()}.}

\item{...}{Custom mappings in name (R version) value (Bioc version) pairs.}

\item{mapp}{Named character vector mapping R versions (names) to
Bioconductor versions (values).
Defaults to \code{map_r_to_bioc()}.}
}
\value{
Returns a Bioconductor version string, or \code{NULL`` if no version yet specified in}mapping_r_to_bioc`.
}
\description{
Given an R version, return a Bioconductor version.
}
