% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{version_cran}
\alias{version_cran}
\alias{map_r_to_snapshot}
\title{Get the designated CRAN date for a version of R}
\usage{
version_cran(r = version_r_major_minor(), map = map_r_to_snapshot())

map_r_to_snapshot(...)
}
\arguments{
\item{r}{The R version (e.g. "3.5").
Defaults to \code{rprofile::version_r_major_minor()}.}

\item{map}{Named character vector mapping R version (names) to snapshot date
(values).
Defaults to \code{rprofile::map_r_to_snapshot()}.}
}
\value{
Returns a date in "YYYY-MM-DD" format.
For unkown R versions, the release date for the current R session version
will be used as a fall back.
}
\description{
This function assigns a date to each version of R, which can be used to set
a versioned CRAN package repository.
}
\examples{
# Get default snapshot date for current R version
rprofile::version_cran()

# Get snapshot intended for version 3.3
rprofile::version_cran("3.3")

# Custom snapshot for R version 3.5.1
custom_snapshots <-
  rprofile::map_r_to_snapshot(
    "3.5.1" = "2100-05-01", # Add a patch level snapshot date
    "3.5"   = "2100-99-01"  # Override existing default
  )

# Use patch level R version and the custom snapshots
rprofile::version_cran(rprofile::version_r(), custom_snapshots)

}
