% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{version_r}
\alias{version_r}
\alias{version_r_major}
\alias{version_r_major_minor}
\title{Granular R versions}
\usage{
version_r()

version_r_major(r = version_r())

version_r_major_minor(r = version_r())
}
\arguments{
\item{r}{Numeric R version as is returned by \code{base::getRversion()}.}
}
\description{
R version utilities. Why? \code{rprofile::version_r_major_minor()} makes the
granularity of the version information clear.
}
