% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{attached_packages}
\alias{attached_packages}
\alias{.SendAttachedPackagesToREnviron}
\alias{.RemoveAttachedPackages}
\alias{.AddAttachedPackagesToDefaultPackages}
\title{Manage attached packages}
\usage{
.SendAttachedPackagesToREnviron()

.RemoveAttachedPackages(attached = search())

.AddAttachedPackagesToDefaultPackages()
}
\arguments{
\item{attached}{A character vector of packages, defaulting to \code{search()}}
}
\description{
Clears out attached packages and properly loads
}
\seealso{
Other Rprofile:
\code{\link[=.AddRprofileOptions]{.AddRprofileOptions()}},
\code{\link[=.AttachDevtools]{.AttachDevtools()}},
\code{\link[=.NiceMessage]{.NiceMessage()}},
\code{\link[=.Search]{.Search()}},
\code{\link[=.UtilMessage]{.UtilMessage()}}
}
\concept{Rprofile}
