% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{attached_packages}
\alias{attached_packages}
\alias{.SendAttachedPackagesToREnviron}
\alias{.RemoveAttachedPackages}
\alias{.AddAttachedPackagesToDefaultPackages}
\title{Manage attached packages}
\usage{
.SendAttachedPackagesToREnviron()

.RemoveAttachedPackages(attached = search())

.AddAttachedPackagesToDefaultPackages()
}
\arguments{
\item{attached}{A character vector of packages, defaulting to \code{search()}}
}
\description{
Clears out attached packages and properly loads
}
\seealso{
Other Rprofile: 
\code{\link{.AddRprofileOptions}()},
\code{\link{.AttachDevtools}()},
\code{\link{.NiceMessage}()},
\code{\link{.Search}()},
\code{\link{.UtilMessage}()},
\code{\link{Rprofile}}
}
\concept{Rprofile}
