% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.AddRprofileOptions}
\alias{.AddRprofileOptions}
\title{Add options}
\usage{
.AddRprofileOptions(...)
}
\arguments{
\item{...}{options to set}
}
\description{
Adds options to \code{rprofile_env}
}
\seealso{
Other Rprofile: 
\code{\link{.AttachDevtools}()},
\code{\link{.NiceMessage}()},
\code{\link{.Rprofile}()},
\code{\link{.UtilMessage}()},
\code{\link{attached_packages}}
}
\concept{Rprofile}
