% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{.NewsUrls}
\alias{.NewsUrls}
\title{Add URLs to NEWS}
\usage{
.NewsUrls(path = ".", ask = interactive(), url = NULL)
}
\arguments{
\item{path}{The path to your package}

\item{ask}{Prompts updating \code{NEWS.md}}

\item{url}{Optional URL to set for bug reports}
}
\value{
Nothing, called for its side effects
}
\description{
Include links to github on \code{NEWS.md}
}
\details{
URLs are detected for GitHub issue.  For example, a line with "Fixes
bug [#123]" will append the URL for the issue number 123.  The repository
is picked up from the \code{DESCRIPTION} file.
}
