% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice-message.R
\name{.NiceMessage}
\alias{.NiceMessage}
\title{Adds a nice message to the start}
\usage{
.NiceMessage(x = 1:2, space = TRUE)
}
\arguments{
\item{x}{A vector of integers to select from.  If the package isn't available
another random message is generated -- or not}

\item{space}{if \code{TRUE} adds another \code{"\\n"} to the output}
}
\description{
Adds a nice message to the start
}
\seealso{
Other Rprofile: 
\code{\link{.AddRprofileOptions}()},
\code{\link{.AttachDevtools}()},
\code{\link{.Rprofile}()},
\code{\link{.Search}()},
\code{\link{.UtilMessage}()},
\code{\link{attached_packages}}
}
\concept{Rprofile}
