% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice-message.R
\name{.NiceMessage}
\alias{.NiceMessage}
\title{Adds a nice message to the start}
\usage{
.NiceMessage(x = 1:2)
}
\arguments{
\item{x}{A vector of integers to select from.  If the package isn't available
another random message is generated -- or not}
}
\description{
Adds a nice message to the start
}
