% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{.Rprofile}
\alias{.Rprofile}
\title{Get .Rprofile}
\usage{
.Rprofile(update = overwrite, overwrite = FALSE)
}
\arguments{
\item{update}{If \code{TRUE} will copy the .Rprofile from the Rprofile package
(defaults to the value of \code{overwrite})}

\item{overwrite}{If \code{TRUE} and \code{update = TRUE} will overwrite .Rprofile if it
exists; ignored if \code{update} is not \code{TRUE}}
}
\description{
Get .Rprofile
}
\seealso{
Other Rprofile: 
\code{\link{.AddRprofileOptions}()},
\code{\link{.AttachDevtools}()},
\code{\link{.NiceMessage}()},
\code{\link{.Search}()},
\code{\link{.UtilMessage}()},
\code{\link{attached_packages}}
}
\concept{Rprofile}
