% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.Search}
\alias{.Search}
\title{Search print}
\usage{
.Search(pattern = ".")
}
\arguments{
\item{pattern}{A regular expression to search for in the search path}
}
\value{
Nothing, called for its side-effects
}
\description{
A more colorful way of printing \code{\link[=search]{search()}}
}
\seealso{
Other Rprofile: 
\code{\link{.AddRprofileOptions}()},
\code{\link{.AttachDevtools}()},
\code{\link{.NiceMessage}()},
\code{\link{.UtilMessage}()},
\code{\link{Rprofile-file}},
\code{\link{attached_packages}}
}
\concept{Rprofile}
