% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{.UtilMessage}
\alias{.UtilMessage}
\title{Utility messages}
\usage{
.UtilMessage(msg)
}
\arguments{
\item{msg}{The message code}
}
\value{
None, called for its side effects
}
\description{
This is exported only for use in \code{.Rprofile}
}
\seealso{
Other Rprofile:
\code{\link[=.AddRprofileOptions]{.AddRprofileOptions()}},
\code{\link[=.AttachDevtools]{.AttachDevtools()}},
\code{\link[=.NiceMessage]{.NiceMessage()}},
\code{\link[=.Search]{.Search()}},
\code{\link{attached_packages}}
}
\concept{Rprofile}
