% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{Rprofile-file}
\alias{Rprofile-file}
\alias{.Rprofile}
\alias{.FindRprofile}
\title{Get .Rprofile}
\usage{
.Rprofile(update = overwrite, overwrite = FALSE, path = .FindRprofile())

.FindRprofile(quiet = FALSE)
}
\arguments{
\item{update}{If \code{TRUE} will copy the .Rprofile from the Rprofile package
(defaults to the value of \code{overwrite})}

\item{overwrite}{If \code{TRUE} and \code{update = TRUE} will overwrite .Rprofile if it
exists; ignored if \code{update} is not \code{TRUE}}

\item{path}{The file location of your \code{.Rprofile}}

\item{quiet}{If \code{TRUE} silences messages, otherwise provides information
about the location of \code{.Rprofile}}
}
\value{
\itemize{
\item \code{.Rprofile()} The \code{.Rprofile()} path, invisibly
}

\itemize{
\item \code{.FindRprofile()} (hopefully) a path to your \code{.Rprofile}
}
}
\description{
Get .Rprofile
}
\seealso{
Other Rprofile: 
\code{\link{.AddRprofileOptions}()},
\code{\link{.AttachDevtools}()},
\code{\link{.NiceMessage}()},
\code{\link{.Search}()},
\code{\link{.UtilMessage}()},
\code{\link{attached_packages}}
}
\concept{Rprofile}
