% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofile-package.R
\docType{package}
\name{Rprofile-package}
\alias{Rprofile}
\alias{Rprofile-package}
\title{Rprofile: Functions to be loaded with Rprofile}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
