% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.rprofile.R
\docType{package}
\name{rprofile}
\alias{rprofile}
\alias{rprofile-package}
\title{Easy manage and provision your Rprofile across machines.}
\description{
Managing your Rprofile when managing multiple machines
or working on a fast-paced team can be challenging.
}
\details{
This package aims to make the process much simpler by
providing a consistent format for manipulating and extending
your Rprofile. Sharing with the team can be accomplished by
referencing other "shared" Rprofiles specified in the same format.
}

