% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_functions}
\alias{set_functions}
\title{Retrieve the hidden environment}
\usage{
set_functions(load_last_error = TRUE, load_last_trace = TRUE,
  load_lsos = TRUE, load_cp = TRUE, load_op = TRUE,
  load_library = TRUE)
}
\arguments{
\item{load_last_error}{Default TRUE}

\item{load_last_trace}{Default TRUE}

\item{load_lsos}{Default TRUE}

\item{load_cp}{Default TRUE}

\item{load_op}{Default TRUE}

\item{load_library}{Default TRUE}
}
\description{
Returns the hidden environment that contains the hidden functions.
}
