% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_startup_info.R
\name{get_active_rproj}
\alias{get_active_rproj}
\alias{set_startup_info}
\title{Customised Startup info}
\usage{
get_active_rproj()

set_startup_info()
}
\description{
Currently prints the numbe of Rsessions running and wifi details.
The \code{get_active_rproj} function needs to be explicitly added to .Rprofile due
to the way RStudio hooks work
}
