% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-options.R
\name{set_startup_options}
\alias{set_startup_options}
\title{Set Nice Startup Options}
\usage{
set_startup_options(
  digits = 4L,
  show.signif.stars = FALSE,
  useFancyQuotes = FALSE,
  width = (cli::console_width() + 17L),
  Ncpus = max(1L, parallel::detectCores() - 1L),
  continue = " ",
  max.print = 100L,
  servr.daemon = TRUE,
  max = 10L,
  mc.cores = max(1L, parallel::detectCores() - 1L),
  warnPartialMatchArgs = TRUE,
  warnPartialMatchDollar = TRUE,
  warnPartialMatchAttr = TRUE,
  scipen = 999L,
  HTTPUserAgent = sprintf("R/\%s R (\%s)", getRversion(), paste(getRversion(),
    R.version$platform, R.version$arch, R.version$os)),
  download.file.extra = sprintf("--header \\"User-Agent: R (\%s)\\"",
    paste(getRversion(), R.version$platform, R.version$arch, R.version$os)),
  setWidthOnResize = TRUE,
  show.error.locations = TRUE,
  check.bounds = TRUE,
  menu.graphics = FALSE,
  error = "rlang",
  ...
)
}
\arguments{
\item{digits}{Default \code{4}}

\item{show.signif.stars}{Default \code{FALSE}}

\item{useFancyQuotes}{Default \code{FALSE}}

\item{width}{Default cli::console_width() + 13L. See details.}

\item{Ncpus}{Default number of CPUs - 1. Used for parallel pkg installs.}

\item{continue}{Default blank space (remove the default +)}

\item{max.print}{Default 100 to avoid blow up}

\item{servr.daemon}{Default \code{TRUE}.  For xaringan presentations}

\item{max}{Default \code{10}. For List printing}

\item{mc.cores}{Default number of CPUs - 1. Used for parallel computing}

\item{warnPartialMatchArgs, warnPartialMatchAttr, warnPartialMatchDollar}{Default \code{TRUE}.
Warn if using partial arguments.}

\item{scipen}{Default \code{999}. Always print out full numbers, i.e. not 1e2}

\item{HTTPUserAgent}{Used by RStudio Package Manager (RSPM).}

\item{download.file.extra}{Used by RSPM for curl/wget installs, e.g. Rscript.}

\item{setWidthOnResize}{If set and TRUE, R run in a terminal using a recent readline
library will set the width option when the terminal is resized.}

\item{show.error.locations}{Show error locations for sourcing}

\item{check.bounds}{logical, defaulting to FALSE. If true, a warning is produced whenever a
vector (atomic or list) is extended, by something like x <- 1:3; x[5] <- 6.}

\item{menu.graphics}{Default \code{FALSE}. Logical: should graphical menus be used if available?}

\item{error}{Default \code{rlang}. If \code{rlang} is installed, then error = rlang::entrace.}

\item{...}{Other arguments passed to \code{options}.}
}
\description{
Sets nicer options. All arguments are passed to the \code{options} function.
This function also sets \code{ipck = TRUE} in \code{rc.settings}.
}
\details{
The \code{width} is only used in an R terminal (not RStudio). However,
fancy prompts that involve colours (such as grey), mean that the column count is off
as grey is \code{grey()(cli::symbol$pointer)} is translated to a 17 character
Unicode string. My slightly hacky solution is to add a line break in the
prompt, than add an additional 17 characters to the width to pad out the width.
}
