% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todos.R
\name{todos}
\alias{todos}
\alias{.Todos}
\alias{.Fixmes}
\title{TODOs and FIXMEs}
\usage{
.Todos(..., .quiet = FALSE, .space = FALSE)

.Fixmes(..., .quiet = FALSE)
}
\arguments{
\item{...}{additional arguments pass to \link[mark:todos]{mark::todos}}

\item{.quiet}{If \code{TRUE} doesn't print result}

\item{.space}{If \code{TRUE} adds another \code{"\\n"} to the output}
}
\value{
See \link[mark:todos]{mark::todos}
}
\description{
Get a list of TODOs and FIXMEs
}
