% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todos.R
\name{todos}
\alias{todos}
\alias{.Todos}
\alias{.Fixmes}
\alias{.TodosHere}
\alias{.FixmesHere}
\title{TODOs and FIXMEs}
\usage{
.Todos(..., .quiet = FALSE, .space = FALSE)

.Fixmes(..., .quiet = FALSE)

.TodosHere(..., .quiet = FALSE)

.FixmesHere(..., .quiet = FALSE)
}
\arguments{
\item{...}{additional arguments pass to \link[mark:todos]{mark::todos}}

\item{.quiet}{If \code{TRUE} doesn't print result}

\item{.space}{If \code{TRUE} adds another \code{"\\n"} to the output}
}
\value{
See \link[mark:todos]{mark::todos}
}
\description{
\code{.Todos()} and \code{.Fixmes()} default to recursively searching the
path \code{"."} and can be changed with setting \code{path} in \code{...}.  \code{.TodosHere()}
and \code{.FixmesHere()} try to search the active file, found with the use of
\code{Rstudio}.
}
\details{
Get a list of TODOs and FIXMEs
}
