% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{version_bioc}
\alias{version_bioc}
\alias{map_r_to_bioc}
\title{Get Bioconductor version}
\usage{
version_bioc(r = version_r_major_minor(), map = map_r_to_bioc())

map_r_to_bioc(...)
}
\arguments{
\item{r}{The R version (e.g. "3.5").
Defaults to \code{\link[rprofile:version_r_major_minor]{rprofile::version_r_major_minor()}}.}

\item{map}{Named character vector mapping R versions (names) to
Bioconductor versions (values).
Defaults to \code{\link[rprofile:map_r_to_bioc]{rprofile::map_r_to_bioc()}}.}

\item{...}{Custom mappings in name (R version) value (Bioc version) pairs.}
}
\value{
Returns a Bioconductor version string, or `NULL`` if no version yet specified
by \code{\link[rprofile:map_r_to_bioc]{rprofile::map_r_to_bioc()}}.
}
\description{
Given an R version, return a Bioconductor version.
}
