% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{map_r_to_snapshot}
\alias{map_r_to_snapshot}
\alias{version_cran}
\title{Get CRAN version}
\usage{
map_r_to_snapshot(...)

version_cran(..., r_name = getOption("rprofile.r_name",
  version_r_major_minor()), latest = getOption("rprofile.latest", FALSE))
}
\arguments{
\item{...}{Custom mappings in name (R version) value (snapshot) pairs.
Overrides default values returned by \code{\link[rprofile:map_r_to_snapshot]{rprofile::map_r_to_snapshot()}}.}

\item{r_name}{An R version (e.g. "3.5") used to select the snapshot.
Defaults to "rprofile.r_name" option if set, else
\code{\link[rprofile:version_r_major_minor]{rprofile::version_r_major_minor()}}.}

\item{latest}{Logical. Use the CRAN mirror without a snapshot date.
Defaults to "rprofile.latest" option if set, else \code{FALSE}.}
}
\value{
Returns a date in "YYYY-MM-DD" format, or "latest".
For unkown R versions, the release date for the current R session version
will be used as a fall back unless \code{latest = TRUE}.
}
\description{
This function assigns a date to each version of R, which can be used to set
a versioned CRAN package repository.
}
\examples{
# Get default snapshot date for current R version
rprofile::version_cran()

# Change the snapshot date for version 3.3
rprofile::version_cran("3.3" = "2100-05-01")

}
